[ENH] Start adding zip file importer (and move XNAT classes)
